# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UNIPROM CRM API
  version: 1.0.0
  description: |
    HTTP API CRM для обработки заявок на изготовление изделий.

    Два взаимозаменяемых префикса с одинаковой серверной проверкой прав и области
    видимости: канонический `/api/v1/...` и контрактный `/api/v1/crm/...`.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Технические эндпоинты дополнительно требуют заголовок `x-integration-token`.

    Создание заявки принимает заголовок `Idempotency-Key`. Изменение ресурса
    требует `lockVersion`; при конфликте возвращается 409 `VERSION_CONFLICT`.
  contact:
    name: UNIPROM
servers:
  - url: http://127.0.0.1:3001
    description: Локальный backend
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: Задачи в области видимости
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: applicationNumber
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Постраничный список задач
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
                  total: { type: integer }
...
  /api/v1/tasks/{taskId}:
    patch:
      tags: [tasks]
      summary: Изменение задачи
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lockVersion]
              properties:
                title: { type: string }
                description: { type: string }
                dueAt: { type: string, format: date-time }
                lockVersion: { type: integer }
      responses:
        '200': { description: Задача изменена }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/tasks/{taskId}/complete:
    post:
      tags: [tasks]
      summary: Закрытие задачи
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: Задача закрыта }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/tasks/{taskId}/cancel:
    post:
      tags: [tasks]
      summary: Отмена задачи
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: Задача отменена }
  /api/v1/applications/{number}/activities:
    get:
      tags: [tasks]
      summary: Коммуникации по заявке
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Коммуникации
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Activity' }
    post:
      tags: [tasks]
      summary: Добавление коммуникации
      description: |
        Для клиентской коммуникации обновляется время последнего контакта и
        снимается блокировка «нет коммерческой активности». Технические события
        помечаются `isCustomerFacing: false` и этот контроль не сбрасывают.
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, body]
              properties:
                type: { type: string }
                body: { type: string }
                isCustomerFacing: { type: boolean, default: true }
      responses:
        '201': { description: Коммуникация создана }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationToken:
      type: apiKey
      in: header
      name: x-integration-token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    Number:
      name: number
      in: path
      required: true
      schema: { type: string }
      description: Номер заявки.
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Conflict:
      description: Конфликт состояния или версии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message: { type: string }
        correlationId: { type: string }
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Task:
      type: object
      properties:
        id: { type: string }
        applicationId: { type: string }
        title: { type: string }
        status: { type: string }
        assigneeId: { type: string, nullable: true }
        dueAt: { type: string, format: date-time, nullable: true }
        lockVersion: { type: integer }
    Activity:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        body: { type: string }
        isCustomerFacing: { type: boolean }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []